requirements.txt
django==3.2
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==6.3.0
djangorestframework~=3.13.1
